+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2020-05-13
+
+### Changed
+
+- Use underscored, lower case column names for ministerial interactions dashboard
+
 ## 2020-05-11
 
 ### Changed